refactor: add logging to export API functions
