+++ lib/contentful/migration.ts
@@ -1,5 +1,6 @@
 import { runMigration } from "contentful-migration"
 
+import { info, success } from "../logger"
 import { MigrationOptions, PendingMigration } from "../types"
 
 export async function runMigrations(
@@ -8,6 +9,8 @@ export async function runMigrations(
 ) {
   const yes = options.yes !== undefined ? options.yes : true
   for (const { filePath } of pendingMigrations) {
+    info(`Deploying migration ${filePath}...`)
     await runMigration({ ...options, filePath, yes })
+    success(`${filePath} migration deployed`)
   }
 }
