+++ lib/createReleaseEnvironment.ts
@@ -9,6 +9,7 @@ import {
   getNextReleaseEnvId,
 } from "./contentful/release"
 import { deployMigrations } from "./deploy"
+import { info, success, warn } from "./logger"
 import { assessPendingMigrations } from "./migrationManagement/migrationState"
 import { MigrationOptions } from "./types"
 
@@ -25,6 +26,12 @@ export async function createReleaseEnvironment({
   ignoreMigrationCheck = false,
   options,
 }: ReleaseOptions) {
+  if (options.environmentId !== "master") {
+    warn(
+      `Releases are intented to work with 'master' aliases. It might no work when running it against another environment`
+    )
+  }
+
   const deployedMigrations = await getDeployedMigrations(options)
   if (!ignoreMigrationCheck) {
     const hasPendingMigrations = await assessPendingMigrations(
@@ -32,6 +39,8 @@ export async function createReleaseEnvironment({
       deployedMigrations
     )
     if (!hasPendingMigrations) {
+      info(`Skipping release since no pending migration was found.`)
+
       return
     }
   }
@@ -45,12 +54,20 @@ export async function createReleaseEnvironment({
   )
   const releaseEnvironmentId = getNextReleaseEnvId(releasePrefix, environments)
   await createEnvironment({ ...options, environmentId: releaseEnvironmentId })
+  info(`Environment ${releaseEnvironmentId} was created.`)
   const deployOptions = { ...options, environmentId: releaseEnvironmentId }
   await deployMigrations({ options: deployOptions, deployedMigrations })
   await updateEnvironmentAlias(
     options.environmentId,
     releaseEnvironmentId,
     options
+  )
+  info(
+    `Alias for the environment ${options.environmentId} was updated to ${releaseEnvironmentId}`
+  )
+
+  success(
+    `Release '${releaseEnvironmentId}' based on the '${options.environmentId}' environment in the space id '${options.spaceId}' was created`
   )
 
   return releaseEnvironmentId
